Organize publisher pipeline and artifacts
