+++ crates/publish/src/vault.rs
@@ -4,4 +4,4 @@ mod scanner;
 pub(crate) use parser::{
     ContentKind, ObsidianFrontMatter, ParsedObsidianFile, parse_obsidian_file,
 };
-pub(crate) use scanner::scan_markdown_files;
+pub(crate) use scanner::{scan_markdown_files, validate_obsidian_dir};
